// src/api/models/BonCommandeCreateRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BonCommandeCreateRequest = {
    numeroCommande: string;
    dateCommande: string;
    dateLivraisonPrevue?: string;
    idFournisseur: string;
    nomFournisseur?: string;
    montantTotal: number;
    montantHT?: number;
    montantTVA?: number;
    devise?: string;
    tauxChange?: number;
    referenceExterne?: string;
    conditionsPaiement?: string;
    delaiLivraison?: number;
    adresseLivraison?: string;
    notes?: string;
    createdBy?: string;
};