// src/api/models/BonCommandeResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BonCommandeResponse = {
    idBonCommande?: string;
    numeroCommande?: string;
    dateCommande?: string;
    dateLivraisonPrevue?: string;
    idFournisseur?: string;
    nomFournisseur?: string;
    montantTotal?: number;
    montantHT?: number;
    montantTVA?: number;
    devise?: string;
    tauxChange?: number;
    statut?: string;
    referenceExterne?: string;
    conditionsPaiement?: string;
    delaiLivraison?: number;
    adresseLivraison?: string;
    notes?: string;
    createdBy?: string;
    createdAt?: string;
    updatedAt?: string;
    validatedAt?: string;
    validatedBy?: string;
};